requirements.txt
pillow-heif==0.13.1
Pillow==10.1.0
customtkinter==5.2.1
